huobi: convert to TS
