--- src/exchanges/huobi-client.js
@@ -1,9 +1,0 @@
-const HuobiBase = require("./huobi-base");
-
-class HuobiClient extends HuobiBase {
-  constructor({ wssPath = "wss://api.huobi.pro/ws", watcherMs } = {}) {
-    super({ name: "Huobi", wssPath, watcherMs });
-  }
-}
-
-module.exports = HuobiClient;
