--- src/exchanges/huobi-japan-client.js
@@ -1,9 +1,0 @@
-const HuobiBase = require("./huobi-base");
-
-class HuobiJapanClient extends HuobiBase {
-  constructor({ wssPath = "wss://api-cloud.huobi.co.jp/ws", watcherMs } = {}) {
-    super({ name: "Huobi Japan", wssPath, watcherMs });
-  }
-}
-
-module.exports = HuobiJapanClient;
