--- src/exchanges/huobi-korea-client.js
@@ -1,9 +1,0 @@
-const HuobiBase = require("./huobi-base");
-
-class HuobiKoreaClient extends HuobiBase {
-  constructor({ wssPath = "wss://api-cloud.huobi.co.kr/ws", watcherMs } = {}) {
-    super({ name: "Huobi Korea", wssPath, watcherMs });
-  }
-}
-
-module.exports = HuobiKoreaClient;
