--- src/exchanges/huobi-russia-client.js
@@ -1,9 +1,0 @@
-const HuobiBase = require("./huobi-base");
-
-class HuobiKoreaClient extends HuobiBase {
-  constructor({ wssPath = "wss://www.huobi.com.ru/-/s/hbcloud/ws", watcherMs } = {}) {
-    super({ name: "Huobi Russia", wssPath, watcherMs });
-  }
-}
-
-module.exports = HuobiKoreaClient;
